Generate an Erlang module from a CSV of message/code mappings.
This replaces the manual maintenance of the mappings in riak_pb_codec
and opens the door for generating the appropriate code for the same
information in other languages.

This also updates the encoding_test to cover the entire message code
space, since we now have gaps in it.

NOTE: I hacked the 'preprocess' callback in the plugin to do the
generation, instead of more appropriately calling it at the beginning
of the compile phase. This deserves more investigation, as I'm
probably doing it wrong.