don't reverse order of records when preparing data for put
Doing so leads to records being laid backwards on the metal
and breaks some optimizations in leveldb which assume the un-reversed
order.

There is a corresponding fix in riak_kv that avoids the second reversal,
via similar prepending new elements to a list of already collected ones,
the net effect of which is the correct order the client eventually sees
on fetch.